l1pf: add FDP & refactor
* add basic FDP counters
* change stride from Block addr to Byte addr
* refactor the code
