+++ src/main/scala/xiangshan/XSTile.scala
@@ -136,7 +136,7 @@ class XSTile()(implicit p: Parameters) extends LazyModule
       misc.l2_binder.get :*= l2.node :*= misc.l1_xbar
       l2.pf_recv_node.map(recv => {
         println("Connecting L1 prefetcher to L2!")
-        recv := core.memBlock.pf_sender_opt.get
+        recv := core.memBlock.l2_pf_sender_opt.get
       })
     case None =>
       val dummyMatch = WireDefault(false.B)
