+++ src/main/scala/xiangshan/mem/prefetch/L1StridePrefetcher.scala
@@ -15,8 +15,8 @@ import scala.collection.SeqLike
 trait HasStridePrefetchHelper extends HasStreamPrefetchHelper {
   val STRIDE_FILTER_SIZE = 6
   val STRIDE_ENTRY_NUM = 10
-  val STRIDE_BITS = 10
-  val STRIDE_VADDR_BITS = 10
+  val STRIDE_BITS = 10 + BLOCK_OFFSET
+  val STRIDE_VADDR_BITS = 10 + BLOCK_OFFSET
   val STRIDE_CONF_BITS = 2
 
   // detail control
@@ -28,154 +28,6 @@ trait HasStridePrefetchHelper extends HasStreamPrefetchHelper {
   val STRIDE_WIDTH_BLOCKS = if(AGGRESIVE_POLICY) STRIDE_LOOK_AHEAD_BLOCKS else 1
   
   def MAX_CONF = (1 << STRIDE_CONF_BITS) - 1
-}
-
-// get prefetch train reqs from `exuParameters.LduCnt` load pipelines (up to `exuParameters.LduCnt`/cycle)
-// filter by cache line address, send out train req to stride (up to 1 req/cycle)
-class StrideTrainFilter()(implicit p: Parameters) extends XSModule with HasStridePrefetchHelper {
-  val io = IO(new Bundle() {
-    val enable = Input(Bool())
-    val flush = Input(Bool())
-    // train input, only from load for now
-    val ld_in = Flipped(Vec(exuParameters.LduCnt, ValidIO(new LdPrefetchTrainBundle())))
-    // filter out
-    val train_req = DecoupledIO(new PrefetchReqBundle())
-  })
-
-  class Ptr(implicit p: Parameters) extends CircularQueuePtr[Ptr]( p => STRIDE_FILTER_SIZE ){}
-  object Ptr {
-    def apply(f: Bool, v: UInt)(implicit p: Parameters): Ptr = {
-      val ptr = Wire(new Ptr)
-      ptr.flag := f
-      ptr.value := v
-      ptr
-    }
-  }
-
-  val entries = RegInit(VecInit(Seq.fill(STRIDE_FILTER_SIZE){ (0.U.asTypeOf(new PrefetchReqBundle())) }))
-  val valids = RegInit(VecInit(Seq.fill(STRIDE_FILTER_SIZE){ (false.B) }))
-
-  // enq
-  val enqLen = exuParameters.LduCnt
-  val enqPtrExt = RegInit(VecInit((0 until enqLen).map(_.U.asTypeOf(new Ptr))))
-  val deqPtrExt = RegInit(0.U.asTypeOf(new Ptr))
-  
-  val deqPtr = WireInit(deqPtrExt.value)
-
-  require(STRIDE_FILTER_SIZE >= enqLen)
-
-  def reorder(source: Vec[ValidIO[LdPrefetchTrainBundle]]): Vec[ValidIO[LdPrefetchTrainBundle]] = {
-    if(source.length == 1) {
-      source
-    }else if(source.length == 2) {
-      val source_v = source.map(_.valid)
-      val res = Wire(source.cloneType)
-      when(PopCount(source_v) === 2.U) {
-        // source 1 is older than source 0
-        val source_1_older = isBefore(source(1).bits.uop.lqIdx, source(0).bits.uop.lqIdx)
-        when(source_1_older) {
-          res(0) := source(1)
-          res(1) := source(0)
-        }.otherwise {
-          res := source
-        }
-      }.otherwise {
-        res := source
-      }
-
-      res
-    }else if(source.length == 3){
-      val res_0_1 = Wire(source.cloneType)
-      val res_1_2 = Wire(source.cloneType)
-      val res = Wire(source.cloneType)
-
-      val tmp = reorder(VecInit(source.slice(0, 2)))
-      res_0_1(0) := tmp(0)
-      res_0_1(1) := tmp(1)
-      res_0_1(2) := source(2)
-      val tmp_1 = reorder(VecInit(res_0_1.slice(1, 3)))
-      res_1_2(0) := res_0_1(0)
-      res_1_2(1) := tmp_1(0)
-      res_1_2(2) := tmp_1(1)
-      val tmp_2 = reorder(VecInit(res_1_2.slice(0, 2)))
-      res(0) := tmp_2(0)
-      res(1) := tmp_2(1)
-      res(2) := res_1_2(2)
-
-      res
-    }else {
-      require(false, "for now, 4 or more sources are invalid")
-      source
-    }
-  }
-
-  val ld_in_reordered = reorder(io.ld_in)
-  val reqs_l = ld_in_reordered.map(_.bits.asPrefetchReqBundle())
-  val reqs_vl = ld_in_reordered.map(_.valid)
-  val needAlloc = Wire(Vec(enqLen, Bool()))
-  val canAlloc = Wire(Vec(enqLen, Bool()))
-
-  for(i <- (0 until enqLen)) {
-    val req = reqs_l(i)
-    val req_v = reqs_vl(i)
-    val index = PopCount(needAlloc.take(i))
-    val allocPtr = enqPtrExt(index)
-    val entry_match = Cat(entries.zip(valids).map {
-      case(e, v) => v && block_hash_tag(e.vaddr) === block_hash_tag(req.vaddr)
-    }).orR
-    val prev_enq_match = if(i == 0) false.B else Cat(reqs_l.zip(reqs_vl).take(i).map {
-      case(pre, pre_v) => pre_v && block_hash_tag(pre.vaddr) === block_hash_tag(req.vaddr)
-    }).orR
-
-    needAlloc(i) := req_v && !entry_match && !prev_enq_match
-    canAlloc(i) := needAlloc(i) && allocPtr >= deqPtrExt && io.enable
-
-    when(canAlloc(i)) {
-      valids(allocPtr.value) := true.B
-      entries(allocPtr.value) := req
-    }
-  }
-  val allocNum = PopCount(canAlloc)
-
-  enqPtrExt.foreach{case x => x := x + allocNum}
-
-  // deq
-  io.train_req.valid := false.B
-  io.train_req.bits := DontCare
-  valids.zip(entries).zipWithIndex.foreach {
-    case((valid, entry), i) => {
-      when(deqPtr === i.U) {
-        io.train_req.valid := valid && io.enable
-        io.train_req.bits := entry
-      }
-    }
-  }
-
-  when(io.train_req.fire) {
-    valids(deqPtr) := false.B
-    deqPtrExt := deqPtrExt + 1.U
-  }
-
-  when(RegNext(io.flush)) {
-    valids.foreach {case valid => valid := false.B}
-    (0 until enqLen).map {case i => enqPtrExt(i) := i.U.asTypeOf(new Ptr)}
-    deqPtrExt := 0.U.asTypeOf(new Ptr)
-  }
-
-  XSPerfAccumulate("stride_train_filter_full", PopCount(valids) === STRIDE_FILTER_SIZE.U)
-  XSPerfAccumulate("stride_train_filter_half", PopCount(valids) >= (STRIDE_FILTER_SIZE / 2).U)
-  XSPerfAccumulate("stride_train_filter_empty", PopCount(valids) === 0.U)
-
-  val raw_enq_pattern = Cat(reqs_vl)
-  val filtered_enq_pattern = Cat(needAlloc)
-  val actual_enq_pattern = Cat(canAlloc)
-  XSPerfAccumulate("stride_train_filter_enq", allocNum > 0.U)
-  XSPerfAccumulate("stride_train_filter_deq", io.train_req.fire)
-  for(i <- 0 until (1 << enqLen)) {
-    XSPerfAccumulate(s"stride_train_filter_raw_enq_pattern_${toBinary(i)}", raw_enq_pattern === i.U)
-    XSPerfAccumulate(s"stride_train_filter_filtered_enq_pattern_${toBinary(i)}", filtered_enq_pattern === i.U)
-    XSPerfAccumulate(s"stride_train_filter_actual_enq_pattern_${toBinary(i)}", actual_enq_pattern === i.U)
-  }
 }
 
 class StrideMetaBundle(implicit p: Parameters) extends XSBundle with HasStridePrefetchHelper {
@@ -192,17 +44,18 @@ class StrideMetaBundle(implicit p: Parameters) extends XSBundle with HasStridePr
   }
 
   def alloc(vaddr: UInt, alloc_hash_pc: UInt) = {
-    pre_vaddr := (block_addr(vaddr))(STRIDE_VADDR_BITS - 1, 0)
+    pre_vaddr := vaddr(STRIDE_VADDR_BITS - 1, 0)
     stride := 0.U
     confidence := 0.U
     hash_pc := alloc_hash_pc
   }
 
   def update(vaddr: UInt, always_update_pre_vaddr: Bool) = {
-    val new_vaddr = (block_addr(vaddr))(STRIDE_VADDR_BITS - 1, 0)
+    val new_vaddr = vaddr(STRIDE_VADDR_BITS - 1, 0)
     val new_stride = new_vaddr - pre_vaddr
+    val new_stride_blk = block_addr(new_stride)
     // NOTE: for now, disable negtive stride
-    val stride_valid = new_stride =/= 0.U && new_stride =/= 1.U && new_stride(STRIDE_VADDR_BITS - 1) === 0.U
+    val stride_valid = new_stride_blk =/= 0.U && new_stride_blk =/= 1.U && new_stride(STRIDE_VADDR_BITS - 1) === 0.U
     val stride_match = new_stride === stride
     val low_confidence = confidence <= 1.U
     val can_send_pf = stride_valid && stride_match && confidence === MAX_CONF.U
@@ -299,9 +152,9 @@ class StrideMetaArray(implicit p: Parameters) extends XSModule with HasStridePre
   val s2_valid = RegNext(s1_valid && s1_can_send_pf)
   val s2_vaddr = RegEnable(s1_vaddr, s1_valid && s1_can_send_pf)
   val s2_stride = RegEnable(s1_stride, s1_valid && s1_can_send_pf)
-  val s2_l1_depth = Cat(s2_stride << l1_stride_ratio, 0.U(BLOCK_OFFSET.W))
+  val s2_l1_depth = s2_stride << l1_stride_ratio
   val s2_l1_pf_vaddr = (s2_vaddr + s2_l1_depth)(VAddrBits - 1, 0)
-  val s2_l2_depth = Cat(s2_stride << l2_stride_ratio, 0.U(BLOCK_OFFSET.W))
+  val s2_l2_depth = s2_stride << l2_stride_ratio
   val s2_l2_pf_vaddr = (s2_vaddr + s2_l2_depth)(VAddrBits - 1, 0)
   val s2_l1_pf_req_bits = (new StreamPrefetchReqBundle).getStreamPrefetchReqBundle(
     vaddr = s2_l1_pf_vaddr,
